+++ config/runtime.exs
@@ -511,8 +511,23 @@ config :lightning, :usage_tracking,
   daily_batch_size: env!("USAGE_TRACKING_DAILY_BATCH_SIZE", :integer, 10)
 
 config :lightning, :kafka_triggers,
-  enabled: env!("KAFKA_TRIGGERS_ENABLED", &Utils.ensure_boolean/1, false),
   duplicate_tracking_retention_seconds:
-    env!("KAFKA_DUPLICATE_TRACKING_RETENTION_SECONDS", :integer, 3600)
+    env!("KAFKA_DUPLICATE_TRACKING_RETENTION_SECONDS", :integer, 3600),
+  enabled: env!("KAFKA_TRIGGERS_ENABLED", &Utils.ensure_boolean/1, false),
+  next_message_candidate_set_delay_milliseconds:
+    env!(
+      "KAFKA_NEXT_MESSAGE_CANDIDATE_SET_DELAY_MILLISECONDS",
+      :integer,
+      250
+    ),
+  no_message_candidate_set_delay_milliseconds:
+    env!(
+      "KAFKA_NO_MESSAGE_CANDIDATE_SET_DELAY_MILLISECONDS",
+      :integer,
+      10000
+    ),
+  number_of_consumers: env!("KAFKA_NUMBER_OF_CONSUMERS", :integer, 1),
+  number_of_message_candidate_set_workers:
+    env!("KAFKA_NUMBER_OF_MESSAGE_CANDIDATE_SET_WORKERS", :integer, 1)
 
 # ==============================================================================